requirements.txt
gradio>=4.0.0
python-docx>=0.8.11
openai>=1.0.0
sentence-transformers>=2.2.0
numpy>=1.21.0
requests>=2.28.0
transformers>=4.21.0
torch>=1.12.0
streamlit>=1.28.0
pandas>=1.5.0
pytest>=7.0.0
